requirements.txt
diffusers==0.31.0
numpy
pillow
torch
transformers